requirements.txt
sqlalchemy==2.0.36
pymysql==1.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
dbt-postgres==1.9.0
pytest==8.3.4
